ast_results/joanpuigsanz_servdroid/android/ServDroid.web/src/org/servDroid/ui/fragment/WebFragment.java
// isComment
package org.servDroid.ui.fragment;

import org.servDroid.helper.IPreferenceHelper;
import org.servDroid.helper.IServiceHelper;
import org.servDroid.server.service.ServerService;
import org.servDroid.ui.activity.ServDroidBaseFragmentActivity;
import org.servDroid.ui.activity.ServDroidBaseFragmentActivity.OnActivityKeyUp;
import org.servDroid.web.R;
import roboguice.inject.InjectView;
import android.app.Activity;
import android.os.Bundle;
import android.view.KeyEvent;
import android.view.LayoutInflater;
import android.view.View;
import android.view.View.OnFocusChangeListener;
import android.view.View.OnKeyListener;
import android.view.ViewGroup;
import android.view.inputmethod.EditorInfo;
import android.webkit.WebView;
import android.webkit.WebViewClient;
import android.widget.EditText;
import android.widget.TextView;
import android.widget.TextView.OnEditorActionListener;
import com.google.inject.Inject;

public class isClassOrIsInterface extends ServDroidBaseFragment implements OnActivityKeyUp {

    @Inject
    IPreferenceHelper isVariable;

    @InjectView(isNameExpr.isFieldAccessExpr.isFieldAccessExpr)
    private WebView isVariable;

    @InjectView(isNameExpr.isFieldAccessExpr.isFieldAccessExpr)
    private EditText isVariable;

    @Inject
    protected IServiceHelper isVariable;

    @Override
    public View isMethod(LayoutInflater isParameter, ViewGroup isParameter, Bundle isParameter) {
        View isVariable = isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr, true);
        return isNameExpr;
    }

    @Override
    public void isMethod(View isParameter, Bundle isParameter) {
        super.isMethod(isNameExpr, isNameExpr);
        isNameExpr.isMethod(new ServDroidWebViewClient());
        isNameExpr.isMethod(new OnKeyListener() {

            public boolean isMethod(View isParameter, int isParameter, KeyEvent isParameter) {
                // isComment
                if ((isNameExpr.isMethod() == isNameExpr.isFieldAccessExpr) && (isNameExpr == isNameExpr.isFieldAccessExpr)) {
                    isNameExpr.isMethod(isNameExpr.isMethod().isMethod());
                    return true;
                }
                return true;
            }
        });
        isNameExpr.isMethod(new OnEditorActionListener() {

            @Override
            public boolean isMethod(TextView isParameter, int isParameter, KeyEvent isParameter) {
                if (isNameExpr == isNameExpr.isFieldAccessExpr) {
                    isNameExpr.isMethod(isNameExpr.isMethod().isMethod());
                }
                return true;
            }
        });
        isNameExpr.isMethod(new OnFocusChangeListener() {

            @Override
            public void isMethod(View isParameter, boolean isParameter) {
                if (isNameExpr) {
                    isNameExpr.isMethod(isNameExpr.isMethod().isMethod());
                }
            }
        });
        isNameExpr.isMethod(true);
        int isVariable = isNameExpr.isMethod();
        if (isNameExpr.isMethod() < isIntegerConstant) {
            isNameExpr = isNameExpr.isFieldAccessExpr;
        }
        isNameExpr.isMethod("isStringConstant" + isNameExpr);
        isNameExpr.isMethod();
    }

    @Override
    public void isMethod() {
        super.isMethod();
        Activity isVariable = isMethod();
        if (isNameExpr instanceof ServDroidBaseFragmentActivity) {
            ((ServDroidBaseFragmentActivity) isNameExpr).isMethod(this);
        }
    }

    @Override
    public void isMethod() {
        super.isMethod();
        Activity isVariable = isMethod();
        if (isNameExpr instanceof ServDroidBaseFragmentActivity) {
            ((ServDroidBaseFragmentActivity) isNameExpr).isMethod(null);
        }
    }

    private class isClassOrIsInterface extends WebViewClient {

        @Override
        public boolean isMethod(WebView isParameter, String isParameter) {
            // isComment
            isNameExpr.isMethod(isNameExpr);
            return true;
        }

        @Override
        public void isMethod(WebView isParameter, String isParameter) {
            super.isMethod(isNameExpr, isNameExpr);
            isNameExpr.isMethod(isNameExpr);
            if (isNameExpr.isMethod()) {
                isNameExpr.isMethod(isNameExpr.isMethod().isMethod());
            }
        }
    }

    @Override
    public boolean isMethod(ServDroidBaseFragmentActivity isParameter, int isParameter, KeyEvent isParameter) {
        if (isNameExpr.isMethod() == isNameExpr.isFieldAccessExpr) {
            switch(isNameExpr) {
                case isNameExpr.isFieldAccessExpr:
                    if (isNameExpr.isMethod() == true) {
                        isNameExpr.isMethod();
                        return true;
                    }
            }
        }
        return true;
    }
}
